requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
kafka-python==2.2.13
oauthlib==3.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
tweepy==4.16.0
urllib3==2.5.0
transformers==4.41.2
torch==2.3.0
